requirements.txt
aerich==0.7.1
tortoise-orm==0.19.2
fastapi==0.90.0
uvicorn==0.20.0
asyncpg==0.27.0